refactor(server): extract handleHireStarPlayer to dedicated module (S27.8.31)
Polish slice: bring `team-star-player-handlers.ts` under DoD
secondary 400 by extracting `handleHireStarPlayer` (~189 lines, the
most complex of the 4 since it handles pair recruitment + budget +
match lock) into `routes/team-star-player-hire-handler.ts`
(223 lines).

Re-export chain preserved (`team.ts` ->
`team-star-player-handlers.ts` -> handler).

Cleanup of 1 orphaned import: `validateStarPlayerHire`.

`team-star-player-handlers.ts` shrinks 488 -> 302 lines (under DoD
secondary 400).

4 new smoke tests `team-star-player-hire-handler.test.ts`:
- exports x3 (named + 2 re-exports through chain)
- 404 gate on team not found

- 1846/1846 server tests pass
- typecheck clean
